move lowercase utils to common
