+++ core/dbt/adapters/cache.py
@@ -16,7 +16,7 @@ from dbt.common.exceptions.cache import (
 )
 from dbt.common.events.functions import fire_event, fire_event_if
 from dbt.adapters.events.types import CacheAction, CacheDumpGraph
-from dbt.utils import lowercase
+from dbt.common.utils.formatting import lowercase
 
 
 def dot_separated(key: _ReferenceKey) -> str:
